+++ tags/release/xlib/envilist.h
@@ -30,22 +30,22 @@ public:
 
   void Clear()  {  Envi.Clear();  }
 
-  inline int Count()    const {  return Envi.Count();  }
+  inline size_t Count() const {  return Envi.Count();  }
   inline bool IsEmpty() const {  return (Envi.Count() == 0);  }
 
   inline TSAtom& GetBase()               const {  return *Base;  }
   inline void SetBase(TSAtom& base)            {  Base = &base;  }
-  inline const olxstr& GetLabel(int ind) const {  return Envi[ind].A()->Label();  }
-  inline TBasicAtomInfo& GetBAI(int ind)       {  return Envi[ind].A()->GetAtomInfo();  }
-  inline TCAtom& GetCAtom(int ind)       const {  return *Envi[ind].A();  }
-  inline const vec3d& GetCrd(int ind)    const {  return Envi[ind].GetC();  }
-  inline const smatd& GetMatrix(int ind) const {  return Envi[ind].GetB();  }
-  void Delete(int i)                           {  Envi.Delete(i);  }
+  inline const olxstr& GetLabel(size_t ind) const {  return Envi[ind].A()->Label();  }
+  inline TBasicAtomInfo& GetBAI(size_t ind)  {  return Envi[ind].A()->GetAtomInfo();  }
+  inline TCAtom& GetCAtom(size_t ind) const {  return *Envi[ind].A();  }
+  inline const vec3d& GetCrd(size_t ind) const {  return Envi[ind].GetC();  }
+  inline const smatd& GetMatrix(size_t ind) const {  return Envi[ind].GetB();  }
+  void Delete(size_t i)  {  Envi.Delete(i);  }
   void SortByDistance()  {
     Envi.QuickSorter.SortMF(Envi, *this, &TAtomEnvi::_SortByDistance);
   }
   void Exclude(TCAtom& ca )                    {
-    for( int i=0; i < Envi.Count(); i++ )
+    for( size_t i=0; i < Envi.Count(); i++ )
       if( Envi[i].GetA() == &ca )  {
         Envi.Delete(i);
         break;
